RatingServer/requirements.txt
argon2-cffi==18.1.0
asgi-redis==1.4.3
asgiref==1.1.2
attrs==18.1.0
autobahn==18.4.1
Automat==0.6.0
awesome-slugify==1.6.5
bleach==2.1.3
boto3==1.6.2
botocore==1.9.23
cassandra-driver==3.14.0
certifi==2018.4.16
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
daphne==1.4.2
Django==1.11.13
django-anymail==2.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.7
django-redis==4.9.0
django-storages==1.6.5
django-taggit==0.22.2
djangorestframework==3.8.2
docutils==0.14
elo==0.1.1
factory-boy==2.11.1
Faker==0.8.13
flake8==3.5.0
flake8-polyfill==1.0.2
gevent==1.2.2
greenlet==0.4.13
gunicorn==19.7.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jmespath==0.9.3
kafka-python==1.4.2
mando==0.6.4
Markdown==2.6.11
mccabe==0.6.1
msgpack-python==0.5.6
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
python-dateutil==2.6.1
pytz==2018.4
radon==2.2.0
raven==6.6.0
redis==2.10.6
regex==2018.2.21
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
sqlparse==0.2.4
text-unidecode==1.2
Twisted==18.4.0
txaio==2.10.0
Unidecode==0.4.21
Unipath==1.1
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
zope.interface==4.5.0
